Session 12/Submission/app/requirements.txt
albumentations==1.3.1
grad-cam==1.4.8
gradio==3.39.0 
numpy== 1.25.0
pillow==9.4.0
pytorch-lightning==2.0.6
pytorch==2.0.1
torch_lr_finder==0.2.1
torchinfo==1.8.0
torchmetrics==0.11.4
torchvision==0.15.2